move Pair to a centrally accessible package
